Make Thunk keep an instance of ThunkInfo directly (NFC)
Previously Thunk was keeping a copy of all the members of ThunkInfo and was creating ThunkInfo on the fly if needed.

This change makes `Thunk` keep an instance of `ThunkInfo` which allows us to move the `ThunkInfo` serialization method into `ThunkInfo` which helps with better separation of concerns and cleaner code.

PiperOrigin-RevId: 782857904