Remove dependency on KernelArguments from CudnnThunk
`KernelArguments` is a type used in code generation and shouldn't leak into thunks.

Instead of taking `KernelArguments` and then converting these into `BufferAllocation::Slice`s in the constructor of `CudnnThunk` we can do this conversion on the call site instead.

To achieve that this change adds a function `KernelArguments::GetArgumentBufferSlices()` for easy acces to the buffer slices.

This also adds a test for `KernelArguments::GetArgumentBufferSlices`.

PiperOrigin-RevId: 782860370